+++ packages/next/src/server-build.ts
@@ -1427,8 +1427,6 @@ export async function serverBuild({
         memory: group.memory,
         runtime: nodeVersion.runtime,
         maxDuration: group.maxDuration,
-        regions: group.regions,
-        functionFailoverRegions: group.functionFailoverRegions,
         supportsCancellation: group.supportsCancellation,
         isStreaming: group.isStreaming,
         nextVersion,